refactor(combat): ♻️ Separate combat functions to their own package
This is needed to prevent the `cyclic import` error.
