+++ api/main.go
@@ -1,8 +1,8 @@
 package main
 
 import (
+	"github.com/PedroChaparro/loomies-backend/combat"
 	"github.com/PedroChaparro/loomies-backend/configuration"
-	"github.com/PedroChaparro/loomies-backend/interfaces"
 	"github.com/PedroChaparro/loomies-backend/routes"
 	"github.com/gin-gonic/gin"
 )
@@ -13,8 +13,8 @@ func main() {
 	routes.SetupRoutes(engine)
 
 	// Setup websocket routes
-	hub := interfaces.WsHub{
-		Combats: make(map[string]*interfaces.WsCombat),
+	hub := combat.WsHub{
+		Combats: make(map[string]*combat.WsCombat),
 	}
 
 	configuration.Globals.WsHub = &hub
